Remove models and stores
