--- src/Models/FactAboutUs.js
@@ -1,4 +1,0 @@
-import { ContentItem } from '@kontent-ai/kontent-delivery';
-
-export class FactAboutUs extends ContentItem {     
-}
