+++ src/Utilities/BrewerFilter.js
@@ -1,0 +1,58 @@
+
+export class Filter {
+  constructor() {
+    this.manufacturers = [];
+    this.priceRanges = [];
+    this.productStatuses = [];
+  }
+
+  matches(brewer) {
+    return this.matchesManufacturers(brewer) && this.matchesPriceRanges(brewer) && this.matchesProductStatuses(brewer);
+  }
+
+  matchesManufacturers(brewer) {
+    if (this.manufacturers.length === 0) {
+      return true;
+    }
+
+    let manufacturerCodenames = brewer.manufacturer.value.map(x => x.codename);
+    return manufacturerCodenames.some(x => this.manufacturers.includes(x));
+  }
+
+  matchesPriceRanges(brewer) {
+    if (this.priceRanges.length === 0) {
+      return true;
+    }
+
+    let price = brewer.price.value;
+
+    return this.priceRanges.some((priceRange) => priceRange.min <= price && price <= priceRange.max);
+  }
+
+  matchesProductStatuses(brewer) {
+    if (this.productStatuses.length === 0) {
+      return true;
+    }
+
+    let statusCodenames = brewer.productStatus.value.map(x => x.codename);
+    return statusCodenames.some((x) => this.productStatuses.includes(x));
+  }
+
+  toggleManufacturer(manufacturer) {
+    let index = this.manufacturers.indexOf(manufacturer);
+
+    if (index < 0) this.manufacturers.push(manufacturer); else this.manufacturers.splice(index, 1);
+  }
+
+  togglePriceRange(priceRange) {
+    let index = this.priceRanges.findIndex((x) => x.min === priceRange.min && x.max === priceRange.max);
+
+    if (index < 0) this.priceRanges.push(priceRange); else this.priceRanges.splice(index, 1);
+  }
+
+  toggleProductStatus(productStatus) {
+    let index = this.productStatuses.indexOf(productStatus);
+
+    if (index < 0) this.productStatuses.push(productStatus); else this.productStatuses.splice(index, 1);
+  }
+}
